doubleの数値がfloatとして処理され誤差が出ていた問題を修正。
git-svn-id: file:///home/moriyoshi/chaperone.e-cell.org/var/opt/subversion/ecell-ide/trunk@3547 fe1fbaf7-ad0a-4b3d-a902-8cb9c273ac6b

+++ ide/EcellUILib/EcellUILib/VariableCountDialog.cs
@@ -319,8 +319,8 @@ namespace Ecell.IDE
             else if (e.ColumnIndex == 1)
             {
                 DataGridViewCell c = (DataGridViewCell)dgv[e.ColumnIndex, e.RowIndex];
-                float dummy;
-                if (!float.TryParse((string)e.FormattedValue, out dummy))
+                double dummy;
+                if (!double.TryParse((string)e.FormattedValue, out dummy))
                 {
                     e.Cancel = true;
                     dgv.CancelEdit();
